// sportpartner-spa-web/src/models/group-detail.model.ts
export class GroupDetailModel {
  groupName: string;
  sportName: string;
  cityName: string;
  id: number;
  isFollowing: boolean;
  totalMember: number;
  totalMessage: number;
  numberNewMembers: number;
  numberRecentMembers: number;
  recentUserIds: number[];
  backgroundImage: string;
  mobileBackgroundImage: string;
  urlPathSport: string;
  urlPathCity: string;
  conversationId: string;
}